extract useCollapsible from OfferList
