+++ src/pages/Browse/TaskDetail/Offers/OfferList.module.scss
@@ -20,33 +20,10 @@
   }
   
   .content{
-    position: relative;
     color: black;
     text-align: left;
-    max-width: 600px;
-    max-height: 600px;
     word-wrap: break-word;
     padding: 8px;
     background-color: rgb(234, 236, 241);
-    transition: max-height 0.3s ease-in-out 0s;
-  }
-
-  .isCollapsed{
-    max-height: 20px;
-    overflow-y: hidden;
-    opacity: 1;
-
-    &::after{
-      position: absolute;
-      right: 0px;
-      bottom: 0px;
-      left: 0px;
-      content: "";
-      display: block;
-      opacity: 1;
-      transition: opacity 0.3s ease-in-out 0s;
-      height: 30px;
-      background: linear-gradient(rgba(255, 255, 255, 0), rgb(246, 248, 253));
-    }
   }
 }
