+++ src/pages/Browse/TaskDetail/Offers/useCollapsible/index.js
@@ -1,0 +1,1 @@
+export { default } from './useCollapsible';
